The which function must return a str object (#26)
* The which function must return a str object

sbank_pipe currently fails on ligolw_add jobs because the path to the file is returned as a str. I'm not sure if I want to do some larger changes here and make the path to this executable be specified in the ini file. In the meantime I patch the existing method.

* Update sbank_pipe

* Update sbank_pipe

* Update sbank_pipe

* Update sbank_pipe